# a7_openapi.yaml
openapi: 3.0.0
info:
 version: '1.0'
 title: 'LBAW Infinity Auction Web API'
 description: 'Web Resources Specification (A7) for Infinity Auction'
servers:
 - url: https://lbaw2271.lbaw.fe.up.pt/
   description: Production server
...
paths:
...
  /users/report/{id}:
    get:
      operationId: R603
      summary: 'R603: Get report form.'
      description: 'Provide user report form UI. Access: USR.'
      tags:
        - 'M06: Reports'
      
      parameters:
        - in: path
          name: reported_user_id
          schema:
            type: integer
            minimum: 1
          required: true

      responses:
        '200':
          description: 'Ok. Show Report form UI.'
...
  /auctions/report/{id}:
    get:
      operationId: R605
      summary: 'R605: Get report form.'
      description: 'Provide auction report form UI. Access: USR.'
      tags:
        - 'M06: Reports'
      
      parameters:
        - in: path
          name: reported_user_id
          schema:
            type: integer
            minimum: 1
          required: true

      responses:
        '200':
          description: 'Ok. Show Report form UI.'
...